+++ pattern/test/src/BasePatternInstructionTest.cpp
@@ -19,7 +19,7 @@ ProjectSpace* BasePatternInstructionTest<T>::space() const {
 }
 
 template<typename T>
-args::patterns* BasePatternInstructionTest<T>::mockPatterns(const int n) {
+args::patterns* BasePatternInstructionTest<T>::mockPatterns(const size_t n) {
     const auto patterns = new PatternsContainerArgument();
     for (size_t i = 0; i < n; ++i) {
         const auto patternSpace = PatternSpace::create();
